non-destructive option

+++ bin/build/htmlify.mjs
@@ -4,11 +4,13 @@ import fs from 'fs';
 import { glob } from 'glob';
 import path from 'path';
 
-function htmlify(filePath, tag) {
+function htmlify(filePath, tag, destroy) {
   const content = fs.readFileSync(filePath);
   const htmlPath = filePath.replace(/\.js$/, '.html');
   fs.writeFileSync(htmlPath, `<${tag}>${content}</${tag}>`);
-  cli.shell.rm(filePath);
+  if (destroy) {
+    cli.shell.rm(filePath);
+  }
   cli.log.info(
     `Htmlified ${cli.colors.url(filePath)} → ${cli.colors.url(
       path.basename(htmlPath)
@@ -27,7 +29,18 @@ function tag(filePath) {
   }
 }
 
-const args = cli.init({ args: { requirePositionals: 2 } });
+const args = cli.init({
+  args: {
+    requirePositionals: true,
+    flags: {
+      destroy: {
+        description:
+          'Delete original file by default (--no-destroy to prevent)',
+        default: true
+      }
+    }
+  }
+});
 glob
   .sync(args.positionals, { cwd: appRootPath.toString(), absolute: true })
-  .map((filePath) => htmlify(filePath, tag(filePath)));
+  .map((filePath) => htmlify(filePath, tag(filePath), args.values.destroy));
